heygen_podcast_bot/requirements.txt
requests
python-dotenv
elevenlabs 